+++ backend/Services/Interfaces/IOrderService.cs
@@ -5,7 +5,7 @@ public interface IOrderService
     Task<List<OrderDto>> GetOrdersAsync(long userId, OrderQueryRequest query);
     Task CancelOrderAsync(long userId, long orderId);
     Task UpdateAddressAsync(long userId, long orderId, long newAddressId);
-    Task<PayAgainResponseDto> PayAgainAsync(int orderId);
+    Task<PayAgainResponseDto> PayAgainAsync(long orderId);
 
     //Admin
     Task<List<OrderDto>> GetAllOrdersAsync(AdminOrderQueryRequest query);
