+++ backend/Services/Interfaces/IRefundService.cs
@@ -1,7 +1,7 @@
 public interface IRefundService
 {
-    Task CreateRefundRequestAsync(int userId, int orderId, CancelPaidOrderRequest request);
+    Task CreateRefundRequestAsync(long userId, long orderId, CancelPaidOrderRequest request);
     Task<List<RefundRequest>> GetAllRefundRequestsAsync();
-    Task ConfirmRefundAsync(int refundId);
-    Task RejectRefundAsync(int refundId, string reason);
+    Task ConfirmRefundAsync(long refundId);
+    Task RejectRefundAsync(long refundId, string reason);
 }
